+++ vkconfig_gui/style.cpp
@@ -21,14 +21,30 @@
 #include "style.h"
 
 #include "../vkconfig_core/configurator.h"
+#include "../vkconfig_core/type_platform.h"
 
+#include <QSysInfo>
 #include <QStyleHints>
 #include <QGuiApplication>
 
 #if QT_VERSION < QT_VERSION_CHECK(6, 5, 0)
 bool IsDarkMode() { return false; }
 #else
-bool IsDarkMode() { return QGuiApplication::styleHints()->colorScheme() == Qt::ColorScheme::Dark; }
+bool IsDarkMode() {
+    if (VKC_PLATFORM == PLATFORM_LINUX) {
+        return false;
+    }
+/*
+    if (VKC_PLATFORM == PLATFORM_LINUX) {
+        std::string OS = QSysInfo::prettyProductName().toStdString();
+        if (OS.find("Ubuntu") == std::string::npos) {
+            return false;
+        }
+    }
+*/
+    Qt::ColorScheme scheme = QGuiApplication::styleHints()->colorScheme();
+    return scheme == Qt::ColorScheme::Dark;
+}
 #endif  // QT_VERSION
 
 QIcon Get(Icon icon) {