+++ vkconfig_gui/tab_preferences.cpp
@@ -35,6 +35,8 @@ TabPreferences::TabPreferences(MainWindow &window, std::shared_ptr<Ui::MainWindo
 #if QT_VERSION < QT_VERSION_CHECK(6, 5, 0)
     this->ui->preferences_theme_mode->setToolTip(
         "Vulkan Configurator must be build with Qt 6.8 or newer to support UI appearance control.");
+#else
+    window.app.setStyle("fusion");
 #endif
 
     this->connect(this->ui->preferences_keep_running, SIGNAL(toggled(bool)), this, SLOT(on_keep_running_toggled(bool)));
@@ -105,7 +107,7 @@ bool TabPreferences::EventFilter(QObject *target, QEvent *event) {
 void TabPreferences::on_theme_mode_changed() {
     const ThemeMode mew_theme_mode = ::IsDarkMode() ? THEME_MODE_FORCE_DARK : THEME_MODE_FORCE_LIGHT;
     if (this->current_theme_mode == mew_theme_mode) {
-        return;
+        // return;
     }
 
     // Configurations
@@ -143,25 +145,25 @@ void TabPreferences::on_theme_mode_changed() {
 
     this->current_theme_mode = mew_theme_mode;
 
-    /*
+    //this->window.update();
+    //this->window.app.styleHints()->setColorScheme(Qt::ColorScheme::Light);
+/*
     #if QT_VERSION >= QT_VERSION_CHECK(6, 8, 0)
-        switch (mode) {
-            case THEME_MODE_USE_DEVICE:
-                this->window.app.styleHints()->unsetColorScheme();
-                break;
+        switch (mew_theme_mode) {
+            default:
             case THEME_MODE_FORCE_LIGHT:
-                if (this->window.app.styleHints()->colorScheme() != Qt::ColorScheme::Light) {
+                //if (this->window.app.styleHints()->colorScheme() != Qt::ColorScheme::Light) {
                     this->window.app.styleHints()->setColorScheme(Qt::ColorScheme::Light);
-                }
+                //}
                 break;
             case THEME_MODE_FORCE_DARK:
-                if (this->window.app.styleHints()->colorScheme() != Qt::ColorScheme::Dark) {
+                //if (this->window.app.styleHints()->colorScheme() != Qt::ColorScheme::Dark) {
                     this->window.app.styleHints()->setColorScheme(Qt::ColorScheme::Dark);
-                }
+                //}
                 break;
         }
     #endif
-    */
+*/
 }
 
 void TabPreferences::on_keep_running_toggled(bool checked) {